update dto
* was updated dto classes

+++ src/main/java/me/kqlqk/behealthy/gateway/dto/conditionService/KcalsInfoDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class KcalsInfoDTO {
     private long id;
-    private short protein;
-    private short fat;
-    private short carb;
+    private int protein;
+    private int fat;
+    private int carb;
 }
